Create README.md
In this project, we are trying to investigate a merchandise dataset provided by Kaggle where “Kaggle” is a franchise for whom we are building a machine learning model that could predict where should Kaggle open their new store (Norway, Finland or Sweden?); additionally, we will also provide what is expected “number of items sold”, and which store is performing well (KaggleMart or KaggleRama?). We use a neural network as a regressor for the time-series dataset. First, we preprocess the dataset based on different feature combinations to perform exploratory analysis and data wrangling. After that, we use the refined dataset for two different learning algorithms. We envisage neural network working as a regressor should outperform the linear regression model for this particular dataset. Finally, we plan to perform hyperparameter tuning and testing and make the final decision along with the required deliverables. The objective is to find out which store can perform better as time forward. Finally using the correct combinations of all operations and configurations we came up with the conclusion that Kaggle Rama is better.

This project is part of the ML Class 2022